requirements.txt
fastapi
uvicorn[standard]
garth
python-dotenv
firebase-admin
pydantic